--- src/main/java/site/gachontable/gachontablebe/global/jwt/exception/ExpiredTokenException.java
@@ -1,10 +1,0 @@
-package site.gachontable.gachontablebe.global.jwt.exception;
-
-import site.gachontable.gachontablebe.global.error.ErrorCode;
-import site.gachontable.gachontablebe.global.error.exception.ServiceException;
-
-public class ExpiredTokenException extends ServiceException {
-    public ExpiredTokenException() {
-        super(ErrorCode.EXPIRED_TOKEN);
-    }
-}
