+++ src/main/java/site/gachontable/infra/security/jwt/filter/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package site.gachontable.gachontablebe.global.jwt;
+package site.gachontable.infra.security.jwt.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
@@ -6,11 +6,11 @@ import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
-import site.gachontable.gachontablebe.global.error.ErrorResponse;
+import site.gachontable.independent.error.ErrorResponse;
 
 import java.io.IOException;
 
-import static site.gachontable.gachontablebe.global.error.ErrorCode.EMPTY_AUTHENTICATION;
+import static site.gachontable.independent.type.ErrorCode.EMPTY_AUTHENTICATION;
 
 @Component
 public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
